Refine core API experimental operators
