+++ README.md
@@ -65,30 +65,15 @@ Backend registration is platform-dependent: `ttx` is available on NPU/ILU/MLU, `
 | Activation | `MojoGelu` | ✅ | ✅ | ✅ | ✅ | TBD | TBD |
 | Activation | `MojoSilu` | ✅ | ✅ | ✅ | ✅ | TBD | TBD |
 | Activation | `MojoSwiGLU` | ✅ | ✅ | ✅ | ✅ | TBD | TBD |
-| Activation | `MojoRotateActivation` | ✅ | TBD | TBD | TBD | TBD | TBD |
 | Attention | `MojoPrefillGQA` | ✅ | ✅ | TBD | TBD | TBD | TBD |
 | Attention | `MojoPagedPrefillGQA` | ✅ | ✅ | ✅ | ✅ | ✅ | ✅ |
-| Attention | `MojoPrefillMLA` | ✅ | TBD | TBD | TBD | TBD | TBD |
-| Attention | `MojoPagedPrefillMLA` | ✅ | TBD | TBD | TBD | TBD | TBD |
-| Attention | `MojoPrefillNSA` | ✅ | TBD | TBD | TBD | TBD | TBD |
-| Attention | `MojoPagedPrefillNSA` | ✅ | TBD | TBD | TBD | TBD | TBD |
 | Attention | `MojoDecodeGQA` | ✅ | TBD | TBD | TBD | TBD | TBD |
 | Attention | `MojoPagedDecodeGQA` | ✅ | ✅ | ✅ | ✅ | ✅ | ✅ |
-| Attention | `MojoDecodeMLA` | ✅ | TBD | TBD | TBD | TBD | TBD |
-| Attention | `MojoPagedDecodeMLA` | ✅ | TBD | TBD | TBD | TBD | TBD |
-| Attention | `MojoDecodeNSA` | ✅ | TBD | TBD | TBD | TBD | TBD |
-| Attention | `MojoPagedDecodeNSA` | ✅ | TBD | TBD | TBD | TBD | TBD |
 | Attention | `MojoSdpa` | ✅ | TBD | ✅ | ✅ | TBD | TBD |
 | Attention | `MojoPagedPrefillSWA` | ✅ | TBD | ✅ | ✅ | ✅ | ✅ |
 | Attention | `MojoPagedDecodeSWA` | ✅ | TBD | ✅ | ✅ | ✅ | ✅ |
 | Attention | `MojoSWA` | ✅ | TBD | ✅ | ✅ | TBD | TBD |
-| Attention | `MojoPagedPrefillGQAWithKVDequant` | ✅ | TBD | TBD | TBD | TBD | ✅ |
-| Attention | `MojoPagedDecodeGQAWithKVDequant` | ✅ | TBD | TBD | TBD | TBD | TBD |
-| Attention | `MojoPagedPrefillSWAWithKVDequant` | ✅ | TBD | TBD | TBD | TBD | ✅ |
-| Attention | `MojoPagedDecodeSWAWithKVDequant` | ✅ | TBD | TBD | TBD | TBD | TBD |
 | KVCache | `MojoStorePagedKVCache` | ✅ | TBD | ✅ | ✅ | TBD | ✅ |
-| KVCache | `MojoStoreMLAKVCache` | ✅ | TBD | TBD | TBD | TBD | TBD |
-| KVCache | `MojoStorePagedMLAKVCache` | ✅ | TBD | TBD | TBD | TBD | TBD |
 | Gemm | `MojoGemm` | ✅ | TBD | TBD | TBD | TBD | TBD |
 | Gemm | `MojoQuantGemm` | ✅ | ✅ | ✅ | TBD | TBD | ✅ |
 | Gemm | `MojoGroupGemm` | ✅ | ✅ | ✅ | ✅ | TBD | ✅ |
@@ -109,32 +94,22 @@ Backend registration is platform-dependent: `ttx` is available on NPU/ILU/MLU, `
 | MoE | `MojoMoE` | ✅ | TBD | TBD | TBD | TBD | TBD |
 | MoE | `MojoMoEGating` | ✅ | TBD | TBD | TBD | TBD | ✅ |
 | MoE | `MojoMoEDispatch` | ✅ | TBD | TBD | TBD | TBD | ✅ |
-| MoE | `MojoMoEInitRoutingDynamicQuant` | ✅ | TBD | TBD | TBD | TBD | TBD |
-| MoE | `MojoFusedSwiGLUMoEScaleDynamicQuantize` | ✅ | TBD | TBD | TBD | TBD | TBD |
 | MoE | `MojoExperts` | ✅ | TBD | TBD | TBD | TBD | TBD |
 | MoE | `MojoMoECombine` | ✅ | TBD | TBD | TBD | TBD | ✅ |
 | MoE | `MojoQuantExperts` | ✅ | TBD | TBD | TBD | TBD | ✅ |
 | MoE | `MojoQuantMoE` | ✅ | TBD | TBD | TBD | TBD | ✅ |
 | Norm | `MojoLayerNorm` | ✅ | TBD | ✅ | ✅ | ✅ | ✅ |
 | Norm | `MojoRMSNorm` | ✅ | ✅ | ✅ | ✅ | TBD | ✅ |
-| Norm | `MojoGroupLayerNorm` | ✅ | TBD | TBD | TBD | TBD | TBD |
 | Norm | `MojoGroupRMSNorm` | ✅ | TBD | TBD | TBD | ✅ | ✅ |
-| Norm | `MojoChannelRMSNorm` | ✅ | TBD | TBD | TBD | TBD | TBD |
 | Norm | `MojoRMSNormQuant` | ✅ | ✅ | TBD | TBD | TBD | TBD |
 | Norm | `MojoLayerNormQuant` | ✅ | ✅ | TBD | TBD | TBD | TBD |
 | Norm | `MojoResidualAddRMSNorm` | ✅ | ✅ | ✅ | ✅ | TBD | ✅ |
 | Norm | `MojoResidualAddLayerNorm` | ✅ | TBD | ✅ | ✅ | TBD | ✅ |
 | Norm | `MojoResidualAddRMSNormQuant` | ✅ | ✅ | TBD | TBD | TBD | TBD |
 | Norm | `MojoResidualAddLayerNormQuant` | ✅ | ✅ | TBD | TBD | TBD | TBD |
-| Norm | `MojoResidualAddNormCast` | ✅ | TBD | TBD | TBD | TBD | TBD |
 | PositionEmb | `MojoRotaryEmbedding` | ✅ | TBD | ✅ | ✅ | TBD | TBD |
-| PositionEmb | `MojoRelativeEmbedding` | ✅ | TBD | TBD | ✅ | TBD | TBD |
 | PositionEmb | `MojoApplyRoPE` | ✅ | ✅ | ✅ | ✅ | ✅ | ✅ |
 | PositionEmb | `MojoApplyVisionRoPE2D` | ✅ | ✅ | ✅ | TBD | TBD | TBD |
-| PositionEmb | `MojoRoPEStoreKV` | ✅ | TBD | TBD | TBD | TBD | TBD |
-| PositionEmb | `MojoNormRoPE` | ✅ | TBD | TBD | TBD | TBD | TBD |
-| PositionEmb | `MojoNormRoPEStoreKV` | ✅ | TBD | TBD | TBD | TBD | TBD |
-| PositionEmb | `MojoGridRoPE` | ✅ | TBD | TBD | TBD | TBD | TBD |
 | PositionEmb | `MojoMRoPE` | ✅ | TBD | ✅ | TBD | TBD | TBD |
 | PositionEmb | `MojoVisionRotaryEmbedding2D` | ✅ | TBD | ✅ | TBD | TBD | TBD |
 | Sampling | `MojoTopPSampling` | ✅ | TBD | ✅ | ✅ | TBD | TBD |
@@ -145,14 +120,34 @@ Backend registration is platform-dependent: `ttx` is available on NPU/ILU/MLU, `
 | Sampling | `MojoTopPFilter` | ✅ | TBD | ✅ | ✅ | TBD | TBD |
 | Convolution | `MojoCausalConv1dUpdateState` | ✅ | TBD | ✅ | ✅ | TBD | TBD |
 | MLP | `MojoSwiGLUMLP` | ✅ | TBD | TBD | TBD | TBD | TBD |
-| Indexer | `MojoLightningIndexer` | ✅ | TBD | ✅ | TBD | TBD | TBD |
 
 ### Experimental Mojo Operator List
 
 | Op Category | Op Name | torch native | torch_npu | ttx_npu | ttx_ilu | ttx_mlu | ixformer |
 | :---------- | :------ | :----------- | :-------- | :------ | :------ | :------ | :------- |
 | Experimental | `MojoQuantBatchGemmReduceSum` | ✅ | ✅ | TBD | ✅ | TBD | TBD |
+| Experimental | `MojoRotateActivation` | ✅ | TBD | TBD | TBD | TBD | TBD |
+| Experimental | `MojoLightningIndexer` | ✅ | TBD | ✅ | TBD | TBD | TBD |
 | Experimental | `MojoIndexer` | ✅ | TBD | ✅ | TBD | TBD | TBD |
+| Experimental | `MojoPrefillMLA` | ✅ | TBD | TBD | TBD | TBD | TBD |
+| Experimental | `MojoPagedPrefillMLA` | ✅ | TBD | TBD | TBD | TBD | TBD |
+| Experimental | `MojoDecodeMLA` | ✅ | TBD | TBD | TBD | TBD | TBD |
+| Experimental | `MojoPagedDecodeMLA` | ✅ | TBD | TBD | TBD | TBD | TBD |
+| Experimental | `MojoPrefillNSA` | ✅ | TBD | TBD | TBD | TBD | TBD |
+| Experimental | `MojoPagedPrefillNSA` | ✅ | TBD | TBD | TBD | TBD | TBD |
+| Experimental | `MojoDecodeNSA` | ✅ | TBD | TBD | TBD | TBD | TBD |
+| Experimental | `MojoPagedDecodeNSA` | ✅ | TBD | TBD | TBD | TBD | TBD |
+| Experimental | `MojoPagedPrefillGQAWithKVDequant` | ✅ | TBD | TBD | TBD | TBD | ✅ |
+| Experimental | `MojoPagedDecodeGQAWithKVDequant` | ✅ | TBD | TBD | TBD | TBD | TBD |
+| Experimental | `MojoPagedPrefillSWAWithKVDequant` | ✅ | TBD | TBD | TBD | TBD | ✅ |
+| Experimental | `MojoPagedDecodeSWAWithKVDequant` | ✅ | TBD | TBD | TBD | TBD | TBD |
+| Experimental | `MojoStorePagedMLAKVCache` | ✅ | TBD | TBD | TBD | TBD | TBD |
+| Experimental | `MojoMoEInitRoutingDynamicQuant` | ✅ | TBD | TBD | TBD | TBD | TBD |
+| Experimental | `MojoFusedSwiGLUMoEScaleDynamicQuantize` | ✅ | TBD | TBD | TBD | TBD | TBD |
+| Experimental | `MojoGroupLayerNorm` | ✅ | TBD | TBD | TBD | TBD | TBD |
+| Experimental | `MojoChannelRMSNorm` | ✅ | TBD | TBD | TBD | TBD | TBD |
+| Experimental | `MojoRelativeEmbedding` | ✅ | TBD | TBD | ✅ | TBD | TBD |
+| Experimental | `MojoGridRoPE` | ✅ | TBD | TBD | TBD | TBD | TBD |
 | Experimental | `MojoStoreLowrank` | ✅ | TBD | ✅ | ✅ | TBD | TBD |
 
 ### Core Mojo Function List
