+++ mojo_opset/backends/ttx/operators/position_embedding.py
@@ -12,10 +12,10 @@ from mojo_opset.backends.ttx.kernels import vision_rope_apply
 from mojo_opset.backends.ttx.kernels import vision_rot_pos_embed
 from mojo_opset.core import MojoApplyRoPE
 from mojo_opset.core import MojoApplyVisionRoPE2D
-from mojo_opset.core import MojoRelativeEmbedding
 from mojo_opset.core import MojoRotaryEmbedding
 from mojo_opset.core import MojoVisionRotaryEmbedding2D
 from mojo_opset.core.operators.position_embedding import MojoMRoPE
+from mojo_opset.experimental import MojoRelativeEmbedding
 
 
 class TTXRelativeEmbedding(MojoRelativeEmbedding):
