+++ mojo_opset/core/operators/position_embedding.py
@@ -1,81 +1,9 @@
-import math
 from typing import Optional
 from typing import Tuple, List
 
 import torch
 
 from ..operator import MojoOperator
-
-
-class MojoRelativeEmbedding(MojoOperator):
-    def __init__(self, num_buckets: int, num_heads: int, bidirectional: bool, max_dist: int = 128):
-        """
-        Initialize T5-style relative position embedding.
-
-        Args:
-            num_buckets (int): Number of relative position buckets.
-            num_heads (int): Attention heads; also the embedding output channels.
-            bidirectional (bool): If True, allocate half buckets for positive direction.
-            max_dist (int, default=128): Maximum distance used in logarithmic bucketing.
-        """
-        super().__init__()
-        if not isinstance(num_buckets, int) or num_buckets <= 0:
-            raise ValueError("num_buckets must be a positive integer")
-        if not isinstance(num_heads, int) or num_heads <= 0:
-            raise ValueError("num_heads must be a positive integer")
-        if not isinstance(bidirectional, bool):
-            raise TypeError("bidirectional must be a bool")
-        if not isinstance(max_dist, int) or max_dist <= 0:
-            raise ValueError("max_dist must be a positive integer")
-        self.num_buckets = num_buckets
-        self.num_heads = num_heads
-        self.bidirectional = bidirectional
-        self.max_dist = max_dist
-        self.embedding = torch.nn.Embedding(num_buckets, num_heads)
-
-    def forward(self, lq: int, lk: int) -> torch.Tensor:
-        """
-        Compute relative position bias tensor for attention.
-
-        Args:
-            lq (int): Length of query sequence (Lq).
-            lk (int): Length of key/value sequence (Lk).
-
-        Returns:
-            torch.Tensor: Bias tensor of shape [1, num_heads, Lq, Lk], dtype follows embedding weights.
-        """
-        if not isinstance(lq, int) or not isinstance(lk, int) or lq <= 0 or lk <= 0:
-            raise ValueError("lq and lk must be positive integers")
-        device = self.embedding.weight.device
-        rel_pos = torch.arange(lk, device=device).unsqueeze(0) - torch.arange(lq, device=device).unsqueeze(1)
-        rel_pos = self._relative_position_bucket(rel_pos)
-        rel_pos_embeds = self.embedding(rel_pos)
-        rel_pos_embeds = rel_pos_embeds.permute(2, 0, 1).unsqueeze(0)
-        return rel_pos_embeds.contiguous()
-
-    def _relative_position_bucket(self, rel_pos: torch.Tensor) -> torch.Tensor:
-        if self.bidirectional:
-            num_buckets = self.num_buckets // 2
-            rel_buckets = (rel_pos > 0).long() * num_buckets
-            rel_pos = torch.abs(rel_pos)
-        else:
-            num_buckets = self.num_buckets
-            rel_buckets = 0
-            rel_pos = -torch.min(rel_pos, torch.zeros_like(rel_pos))
-
-        max_exact = num_buckets // 2
-        rel_pos_large = (
-            max_exact
-            + (
-                torch.log(rel_pos.float() / max_exact) / math.log(self.max_dist / max_exact) * (num_buckets - max_exact)
-            ).long()
-        )
-        rel_pos_large = torch.min(rel_pos_large, torch.full_like(rel_pos_large, num_buckets - 1))
-        rel_buckets += torch.where(rel_pos < max_exact, rel_pos, rel_pos_large)
-        return rel_buckets
-
-    def extra_repr(self) -> str:
-        return f"{self.num_buckets=}, {self.num_heads=}, {self.bidirectional=}, {self.max_dist=}".replace("self.", "")
 
 
 class MojoRotaryEmbedding(MojoOperator):
@@ -245,59 +173,6 @@ class MojoApplyRoPE(MojoOperator):
             cos = cos.unsqueeze(-2)
             sin = sin.unsqueeze(-2)
         return self._apply_rope(q, k, cos, sin)
-
-
-class MojoRoPEStoreKV(MojoOperator):
-    pass
-
-
-class MojoNormRoPE(MojoOperator):
-    pass
-
-
-class MojoNormRoPEStoreKV(MojoOperator):
-    pass
-
-
-class MojoGridRoPE(MojoOperator):
-    def __init__(self):
-        super().__init__()
-
-    def forward(
-        self,
-        x: torch.Tensor,
-        grid_sizes: torch.Tensor,
-        freqs_list: List[torch.Tensor],
-    ) -> torch.Tensor:
-        """
-        Apply 3D grid rotary position embeddings (RoPE) over (F, H, W) axes using
-        precomputed per-sample frequency tensors.
-
-        Args:
-            x (torch.Tensor): [B, L, N, D]; D must be even (paired into complex components).
-            grid_sizes (torch.Tensor): [B, 3] per-sample (F, H, W); seq_len = F*H*W.
-            freqs_list (List[torch.Tensor]): length-B list; each item is a complex unit-phase tensor
-                of shape [seq_len, 1, D/2], broadcastable to [seq_len, N, D/2].
-
-        Returns:
-            torch.Tensor: Same shape as `x`. Per sample, the first F*H*W tokens are rotated;
-                remaining padding tokens are preserved. Output dtype matches input.
-        """
-        assert x.dim() == 4, "x must be 4D: [B, L, N, D]"
-        assert x.size(-1) % 2 == 0, "D must be even for complex pairing"
-        assert grid_sizes.dim() == 2 and grid_sizes.size(1) == 3, "grid_sizes must be [B, 3]"
-
-        n = x.size(2)
-        output = []
-        for i, (f, h, w) in enumerate(grid_sizes.tolist()):
-            seq_len = f * h * w
-            x_i = torch.view_as_complex(x[i, :seq_len].to(torch.float32).reshape(seq_len, n, -1, 2))
-            freqs_i = freqs_list[i]
-            x_i = torch.view_as_real(x_i * freqs_i).flatten(2)
-            x_i = torch.cat([x_i, x[i, seq_len:]])
-            output.append(x_i)
-        y = torch.stack(output)
-        return y.type_as(x)
 
 
 class MojoMRoPE(MojoOperator):
