+++ mojo_opset/experimental/__init__.py
@@ -5,14 +5,56 @@ Once we find the operators of contrib become more and more stable in community, 
 
 from .functions.diffusion_attention import MojoDiffusionAttentionFunction
 from .functions.diffusion_attention import mojo_diffusion_attention
+from .operators.activation import MojoRotateActivation
+from .operators.attention import MojoDecodeMLA
+from .operators.attention import MojoDecodeNSA
+from .operators.attention import MojoPagedDecodeMLA
+from .operators.attention import MojoPagedDecodeNSA
+from .operators.attention import MojoPagedDecodeGQAWithKVDequant
+from .operators.attention import MojoPagedDecodeSWAWithKVDequant
+from .operators.attention import MojoPagedPrefillGQAWithKVDequant
+from .operators.attention import MojoPagedPrefillMLA
+from .operators.attention import MojoPagedPrefillNSA
+from .operators.attention import MojoPagedPrefillSWAWithKVDequant
+from .operators.attention import MojoPrefillMLA
+from .operators.attention import MojoPrefillNSA
 from .operators.gemm import MojoQuantBatchGemmReduceSum
 from .operators.indexer import MojoIndexer
+from .operators.indexer import MojoLightningIndexer
+from .operators.kv_cache import MojoStorePagedMLAKVCache
+from .operators.moe import MojoFusedSwiGLUMoEScaleDynamicQuantize
+from .operators.moe import MojoMoEInitRoutingDynamicQuant
+from .operators.normalization import MojoChannelRMSNorm
+from .operators.normalization import MojoGroupLayerNorm
+from .operators.position_embedding import MojoGridRoPE
+from .operators.position_embedding import MojoRelativeEmbedding
 from .operators.store_lowrank import MojoStoreLowrank
 
 __all__ = [
     "MojoDiffusionAttentionFunction",
     "mojo_diffusion_attention",
+    "MojoRotateActivation",
     "MojoQuantBatchGemmReduceSum",
+    "MojoLightningIndexer",
+    "MojoPrefillMLA",
+    "MojoPagedPrefillMLA",
+    "MojoDecodeMLA",
+    "MojoPagedDecodeMLA",
+    "MojoPrefillNSA",
+    "MojoPagedPrefillNSA",
+    "MojoDecodeNSA",
+    "MojoPagedDecodeNSA",
+    "MojoPagedPrefillGQAWithKVDequant",
+    "MojoPagedDecodeGQAWithKVDequant",
+    "MojoPagedPrefillSWAWithKVDequant",
+    "MojoPagedDecodeSWAWithKVDequant",
+    "MojoStorePagedMLAKVCache",
+    "MojoMoEInitRoutingDynamicQuant",
+    "MojoFusedSwiGLUMoEScaleDynamicQuantize",
+    "MojoGroupLayerNorm",
+    "MojoChannelRMSNorm",
+    "MojoRelativeEmbedding",
+    "MojoGridRoPE",
     "MojoStoreLowrank",
     "MojoIndexer",
 ]
