+++ mojo_opset/experimental/operators/__init__.py
@@ -1,9 +1,51 @@
+from .activation import MojoRotateActivation
+from .attention import MojoDecodeMLA
+from .attention import MojoDecodeNSA
+from .attention import MojoPagedDecodeMLA
+from .attention import MojoPagedDecodeNSA
+from .attention import MojoPagedDecodeGQAWithKVDequant
+from .attention import MojoPagedDecodeSWAWithKVDequant
+from .attention import MojoPagedPrefillGQAWithKVDequant
+from .attention import MojoPagedPrefillMLA
+from .attention import MojoPagedPrefillNSA
+from .attention import MojoPagedPrefillSWAWithKVDequant
+from .attention import MojoPrefillMLA
+from .attention import MojoPrefillNSA
 from .gemm import MojoQuantBatchGemmReduceSum
 from .indexer import MojoIndexer
+from .indexer import MojoLightningIndexer
+from .kv_cache import MojoStorePagedMLAKVCache
+from .moe import MojoFusedSwiGLUMoEScaleDynamicQuantize
+from .moe import MojoMoEInitRoutingDynamicQuant
+from .normalization import MojoChannelRMSNorm
+from .normalization import MojoGroupLayerNorm
+from .position_embedding import MojoGridRoPE
+from .position_embedding import MojoRelativeEmbedding
 from .store_lowrank import MojoStoreLowrank
 
 __all__ = [
+    "MojoRotateActivation",
     "MojoIndexer",
+    "MojoLightningIndexer",
+    "MojoPrefillMLA",
+    "MojoPagedPrefillMLA",
+    "MojoDecodeMLA",
+    "MojoPagedDecodeMLA",
+    "MojoPrefillNSA",
+    "MojoPagedPrefillNSA",
+    "MojoDecodeNSA",
+    "MojoPagedDecodeNSA",
+    "MojoPagedPrefillGQAWithKVDequant",
+    "MojoPagedDecodeGQAWithKVDequant",
+    "MojoPagedPrefillSWAWithKVDequant",
+    "MojoPagedDecodeSWAWithKVDequant",
+    "MojoStorePagedMLAKVCache",
+    "MojoMoEInitRoutingDynamicQuant",
+    "MojoFusedSwiGLUMoEScaleDynamicQuantize",
+    "MojoGroupLayerNorm",
+    "MojoChannelRMSNorm",
+    "MojoRelativeEmbedding",
+    "MojoGridRoPE",
     "MojoStoreLowrank",
     "MojoQuantBatchGemmReduceSum",
 ]
